Moved levelEditor workshop files into their own subfolder + Fixed levelSelector not getting properly initialised
